Create How to add a file to your repository
Files in GitHub can do a handful of things, but the main purpose of files is to store data and information about your project. It's worth knowing in order to add a file to a repository that you must first have minimum Write access within the repository you want to add a file.

Congratulations, you just created a new file in your repository! You have also created a new branch and made a commit.